wip: calculate questions on initial sync
