+++ server/bleep/src/background/sync.rs
@@ -1,4 +1,3 @@
-use anyhow::{bail, Context};
 use either::Either;
 use tokio::sync::OwnedSemaphorePermit;
 use tracing::{debug, error, info, warn};
@@ -8,7 +7,6 @@ use crate::{
     indexes,
     remotes::RemoteError,
     repo::{Backend, RepoError, RepoMetadata, RepoRef, Repository, SyncStatus},
-    state::RepositoryPool,
     Application,
 };
 
@@ -180,20 +178,25 @@ impl SyncHandle {
             return Err(SyncError::Cancelled);
         }
 
-        // TODO: this should only run once
-        let tutorial_questions = {
+        // Can we unwrap here?
+        let repository = repo_pool
+            .read_async(&self.reporef, |_k, v| v.clone())
+            .await
+            .unwrap();
+
+        let tutorial_questions = if repository.last_index_unix_secs == 0 {
             let db = self.app.sql.clone();
             let llm_gateway = self.app.llm_gateway_client();
             let repo_pool = self.app.repo_pool.clone();
             let reporef = self.reporef.clone();
 
-            tokio::task::spawn(generate_tutorial_questions(
-                db,
-                llm_gateway,
-                repo_pool,
-                reporef,
+            Some(tokio::task::spawn(
+                crate::commits::generate_tutorial_questions(db, llm_gateway, repo_pool, reporef),
             ))
+        } else {
+            None
         };
+
         let indexed = self.index().await;
         let status = match indexed {
             Ok(Either::Left(status)) => Some(status),
@@ -203,8 +206,10 @@ impl SyncHandle {
                     repo.sync_done_with(self.new_branch_filters.as_ref(), state)
                 });
 
-                if let Err(err) = tutorial_questions.await {
-                    error!(?err, "failed to generate tutorial questions");
+                if let Some(tutorial_questions) = tutorial_questions {
+                    if let Err(err) = tutorial_questions.await {
+                        error!(?err, "failed to generate tutorial questions");
+                    }
                 }
 
                 // technically `sync_done_with` does this, but we want to send notifications
@@ -466,64 +471,4 @@ impl SyncHandle {
             repo.expect("repo was already deleted")
         }
     }
-}
-
-async fn generate_tutorial_questions(
-    db: crate::db::SqlDb,
-    llm_gateway: anyhow::Result<crate::llm_gateway::Client>,
-    repo_pool: RepositoryPool,
-    reporef: RepoRef,
-) -> anyhow::Result<()> {
-    let repo_str = reporef.to_string();
-    let rows = sqlx::query! {
-        "SELECT * FROM tutorial_questions \
-         WHERE repo_ref = ?",
-        repo_str,
-    }
-    .fetch_all(db.as_ref())
-    .await?;
-
-    if !rows.is_empty() {
-        debug!(%reporef, "skipping tutorial questions, already have some");
-        return Ok(());
-    }
-
-    debug!(%reporef, "generating tutorial questions");
-    let Ok(llm_gateway) = llm_gateway
-    else {
-	bail!("badly configured llm gw");
-    };
-
-    // Due to `Send` issues on the gix side, we need to split this off quite brutally.
-    let latest_commits = {
-        let reporef = reporef.clone();
-        tokio::task::spawn_blocking(|| crate::commits::latest_commits(repo_pool, reporef, None))
-            .await
-            .context("threads error")??
-    };
-
-    error!("About to expand commits to suggestions");
-    let suggestions =
-        crate::commits::expand_commits_to_suggestions(latest_commits, &llm_gateway).await?;
-    error!("{:?}", &suggestions);
-
-    debug!(%reporef, count=suggestions.len(), "found suggestions");
-
-    let mut tx = db.begin().await?;
-    for q in suggestions {
-        _ = sqlx::query!(
-            "INSERT INTO tutorial_questions (question, tag, repo_ref) \
-             VALUES (?, ?, ?)",
-            q.question,
-            q.tag,
-            repo_str,
-        )
-        .execute(&mut tx)
-        .await?;
-    }
-
-    tx.commit().await?;
-
-    debug!(%reporef, "suggestions committed");
-    Ok(())
 }
